Add a Dockerfile that allows efficient local builds
The Dockerfile essentially builds a docker image by downloading
moergo-sc/zmk and the current main branch of
moergo-sc/glove80-zmk-config, and then building the latter to warm up
the image's nix store.

When running the image, very little gets downloaded, and the image
just builds glove80.uf2 and copies it back to the current
directory.

The provided `build.sh` script both builds and runs the docker image,
calling docker with the appropriate arguments.